Add method in ActionFactory
