+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/action/Action.java
@@ -18,100 +18,9 @@
 
 package org.apache.paimon.flink.action;
 
-import org.apache.paimon.catalog.CatalogUtils;
-import org.apache.paimon.utils.Preconditions;
-
-import org.apache.flink.api.java.tuple.Tuple3;
-import org.apache.flink.api.java.utils.MultipleParameterTool;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
 /** Abstract class for Flink actions. */
 public interface Action {
 
     /** The execution method of the action. */
     void run() throws Exception;
-
-    static Tuple3<String, String, String> getTablePath(MultipleParameterTool params) {
-        String warehouse = params.get("warehouse");
-        String database = params.get("database");
-        String table = params.get("table");
-        String path = params.get("path");
-
-        Tuple3<String, String, String> tablePath = null;
-        int count = 0;
-        if (warehouse != null || database != null || table != null) {
-            if (warehouse == null || database == null || table == null) {
-                throw new IllegalArgumentException(
-                        "Warehouse, database and table must be specified all at once to specify a table.");
-            }
-            tablePath = Tuple3.of(warehouse, database, table);
-            count++;
-        }
-        if (path != null) {
-            tablePath =
-                    Tuple3.of(
-                            CatalogUtils.warehouse(path),
-                            CatalogUtils.database(path),
-                            CatalogUtils.table(path));
-            count++;
-        }
-
-        if (count != 1) {
-            throw new IllegalArgumentException(
-                    "Please specify either \"warehouse, database and table\" or \"path\".");
-        }
-
-        return tablePath;
-    }
-
-    static List<Map<String, String>> getPartitions(MultipleParameterTool params) {
-        List<Map<String, String>> partitions = new ArrayList<>();
-        for (String partition : params.getMultiParameter("partition")) {
-            Map<String, String> kvs = parseCommaSeparatedKeyValues(partition);
-            partitions.add(kvs);
-        }
-
-        return partitions;
-    }
-
-    static Map<String, String> parseCommaSeparatedKeyValues(String keyValues) {
-        Map<String, String> kvs = new HashMap<>();
-        for (String kvString : keyValues.split(",")) {
-            parseKeyValueString(kvs, kvString);
-        }
-        return kvs;
-    }
-
-    static Map<String, String> optionalConfigMap(MultipleParameterTool params, String key) {
-        if (!params.has(key)) {
-            return Collections.emptyMap();
-        }
-
-        Map<String, String> config = new HashMap<>();
-        for (String kvString : params.getMultiParameter(key)) {
-            parseKeyValueString(config, kvString);
-        }
-        return config;
-    }
-
-    static void parseKeyValueString(Map<String, String> map, String kvString) {
-        String[] kv = kvString.split("=");
-        if (kv.length != 2) {
-            throw new IllegalArgumentException(
-                    String.format(
-                            "Invalid key-value string '%s'. Please use format 'key=value'",
-                            kvString));
-        }
-        map.put(kv[0].trim(), kv[1].trim());
-    }
-
-    static void checkRequiredArgument(MultipleParameterTool params, String key) {
-        Preconditions.checkArgument(
-                params.has(key), "Argument '%s' is required. Run '<action> --help' for help.", key);
-    }
 }
