+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/action/DeleteActionFactory.java
@@ -24,9 +24,6 @@ import org.apache.flink.api.java.utils.MultipleParameterTool;
 import java.util.Map;
 import java.util.Optional;
 
-import static org.apache.paimon.flink.action.Action.getTablePath;
-import static org.apache.paimon.flink.action.Action.optionalConfigMap;
-
 /** Factory to create {@link DeleteAction}. */
 public class DeleteActionFactory implements ActionFactory {
